More updates to build portion of documentation. Verified that it works, as written, for a mac, at least for a "clean" machine and up through the parallel build (did not check Trilinos build yet).
Note that some changes have simplified old version of openmpi build / install process. These are reflected in the
documentation but would require the following lines in the cmake build script (for mac ONLY) to be changed from:

  -D CMAKE_CXX_COMPILER=openmpicxx \
  -D CMAKE_C_COMPILER=openmpicc \
  -D CMAKE_Fortran_COMPILER=openmpif90 \

to ...

  -D CMAKE_CXX_COMPILER=mpicxx \
  -D CMAKE_C_COMPILER=mpicc \
  -D CMAKE_Fortran_COMPILER=mpif90 \

I have NOT committed the change to the cmake script for macs yet, so the documentation is inconsistent in that respect.

Still deciding on whether or not the build process should be changed (e.g. by sym linking the openmpi calls to mpi
equivalents) or if the cmake script should be changed.





git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/cism2_documentation@1561 42a2fd4b-9c52-44de-a8a9-1772c67210cd