Checking in some edits to the first part of Chapter 5.  The equations are now numbered (and labeled as needed).
I inserted a few text comments that Matt or Steve may want to look at.

I stopped partway through section 3 on boundary conditions.  This section seems to have been written to describe
a finite-difference discretization of BCs as in Glam.  Since these equations aren't used in the Glissade solver,
I don't think they need to be included here.  So the section probably needs to be rewritten, but I don't
want to take that on right now.



git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/cism2_documentation@1686 42a2fd4b-9c52-44de-a8a9-1772c67210cd