DOCUMENTATION
Adding description of the 2 new basal friction laws.  The description and equations are
in the Specified Basal Traction section (5.3.2) of Chapter 5: Introduction to 
Higher-Order Ice Dynamics.  The config settings are added to the existing tables in 
the Running CISM chapter.
I also added equations numbers to all the equations in the higher-order-model-bcs.tex 
file (the multi-lined equations were missing numbers).


git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/cism2_documentation@1794 42a2fd4b-9c52-44de-a8a9-1772c67210cd