DOCUMENTATION: Updating notation for consistency between chapters.
Edits to chapters: Intro to Modeling, Glide Dynamics, Intro to HO modeling, Glissade Dynamics

* generic source term H has become R in Intro to Modeling Chapter
* thickness has become H throughout (was h in some places)
* surface mass balance has become B_s throughout (was various)
* basal mass balance has become B_b throughout (was various)
* combined surface and basal mass balance has become B throughout (was various)
* bed elevation has become b throughout (was various)
* bedrock elevation has become b_r

I have added an Appendix with notation for a handful of commonly used variables.  This is pretty barebones, but at least codifies the usage that I carefully checked.

One remaining conflict I noticed is the use of B for the rate factor in Eq. 5.11 (and elsewhere in that chapter) conflicts with B meaning combined surface and basal mass balance elsewhere.  I think the usage of B there is described carefully enough that this is ok.  Maybe B should become B_c to eliminate this potential confusion.

Also added missing equation numbers to some sections in the Intro to HO chapter.


git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/cism2_documentation@1820 42a2fd4b-9c52-44de-a8a9-1772c67210cd